Add admin docs and cleanup comments

+++ gamemode/core/libraries/admin.lua
@@ -3,7 +3,12 @@ lia.admin.bans = lia.admin.bans or {}
 lia.admin.groups = lia.admin.groups or {}
 lia.admin.banList = lia.admin.banList or {}
 lia.admin.privileges = lia.admin.privileges or {}
+
+function lia.admin.isDisabled()
+    return sam ~= nil or serverguard ~= nil
+end
 function lia.admin.load()
+    if lia.admin.isDisabled() then return end
     lia.admin.groups = lia.data.get("admin_groups", {}, true, true)
     for name, priv in pairs(CAMI.GetPrivileges() or {}) do
         lia.admin.privileges[name] = priv
@@ -17,6 +22,7 @@ function lia.admin.load()
 end
 
 function lia.admin.createGroup(groupName, info)
+    if lia.admin.isDisabled() then return end
     if lia.admin.groups[groupName] then
         Error("[Lilia Administration] This usergroup already exists!\n")
         return
@@ -27,11 +33,13 @@ function lia.admin.createGroup(groupName, info)
 end
 
 function lia.admin.registerPrivilege(privilege)
+    if lia.admin.isDisabled() then return end
     if not privilege or not privilege.Name then return end
     lia.admin.privileges[privilege.Name] = privilege
 end
 
 function lia.admin.removeGroup(groupName)
+    if lia.admin.isDisabled() then return end
     if not lia.admin.groups[groupName] then
         Error("[Lilia Administration] This usergroup doesn't exist!\n")
         return
@@ -43,6 +51,7 @@ end
 
 if SERVER then
     function lia.admin.addPermission(groupName, permission)
+        if lia.admin.isDisabled() then return end
         if not lia.admin.groups[groupName] then
             Error("[Lilia Administration] This usergroup doesn't exist!\n")
             return
@@ -53,6 +62,7 @@ if SERVER then
     end
 
     function lia.admin.removePermission(groupName, permission)
+        if lia.admin.isDisabled() then return end
         if not lia.admin.groups[groupName] then
             Error("[Lilia Administration] This usergroup doesn't exist!\n")
             return
@@ -63,6 +73,7 @@ if SERVER then
     end
 
     function lia.admin.save(network)
+        if lia.admin.isDisabled() then return end
         lia.data.set("admin_groups", lia.admin.groups, true, true)
         if network then
             net.Start("lilia_updateAdminGroups")
@@ -72,11 +83,13 @@ if SERVER then
     end
 
     function lia.admin.setPlayerGroup(ply, usergroup)
+        if lia.admin.isDisabled() then return end
         ply:SetUserGroup(usergroup)
         lia.db.query(Format("UPDATE lia_players SET _userGroup = '%s' WHERE _steamID = %s", lia.db.escape(usergroup), ply:SteamID64()))
     end
 
     function lia.admin.addBan(steamid, reason, duration)
+        if lia.admin.isDisabled() then return end
         if not steamid then Error("[Lilia Administration] lia.admin.addBan: no steam id specified!") end
         local banStart = os.time()
         lia.admin.banList[steamid] = {
@@ -94,32 +107,43 @@ if SERVER then
     end
 
     function lia.admin.removeBan(steamid)
+        if lia.admin.isDisabled() then return end
         if not steamid then Error("[Lilia Administration] lia.admin.removeBan: no steam id specified!") end
         lia.admin.banList[steamid] = nil
         lia.db.query(Format("DELETE FROM lia_bans WHERE _steamID = '%s'", lia.db.escape(steamid)), function() MsgC(Color(0, 200, 0), "[Lilia Administration] Ban removed.\n") end)
     end
 
     function lia.admin.isBanned(steamid)
+        if lia.admin.isDisabled() then return false end
         return lia.admin.banList[steamid] or false
     end
 
     function lia.admin.hasBanExpired(steamid)
+        if lia.admin.isDisabled() then return true end
         local ban = lia.admin.banList[steamid]
         if not ban then return true end
         if ban.duration == 0 then return false end
         return ban.start + ban.duration <= os.time()
     end
 
-    hook.Add("InitPostEntity", "lia_LoadAdmin", function() lia.admin.load() end)
-    hook.Add("ShutDown", "lia_SaveAdmin", function() lia.admin.save() end)
+    hook.Add("InitPostEntity", "lia_LoadAdmin", function()
+        if lia.admin.isDisabled() then return end
+        lia.admin.load()
+    end)
+    hook.Add("ShutDown", "lia_SaveAdmin", function()
+        if lia.admin.isDisabled() then return end
+        lia.admin.save()
+    end)
 end
 
 hook.Add("PlayerAuthed", "lia_SetUserGroup", function(ply, steamID)
+    if lia.admin.isDisabled() then return end
     local steam64 = util.SteamIDTo64(steamID)
     lia.db.query(Format("SELECT _userGroup FROM lia_players WHERE _steamID = %s", steam64), function(data) if istable(data) and data[1] then ply:SetUserGroup(data[1]._userGroup) end end)
 end)
 
 hook.Add("OnDatabaseLoaded", "lia_LoadBans", function()
+    if lia.admin.isDisabled() then return end
     lia.db.query("SELECT * FROM lia_bans", function(data)
         if istable(data) then
             local bans = {}
@@ -137,6 +161,7 @@ hook.Add("OnDatabaseLoaded", "lia_LoadBans", function()
 end)
 
 concommand.Add("plysetgroup", function(ply, cmd, args)
+    if lia.admin.isDisabled() then return end
     if not IsValid(ply) then
         local target = lia.util.findPlayer(args[1])
         if IsValid(target) then
